Added support for the old queue behavior - limit the max queue size.
This fixed the following issues:
Zendesk#323 - MoH from Park Server unintelligible after 15 seconds
PT44534121 - Call Park music will stop playing after 15 secs

The old behavior, have a limited queue,  was needed because media components
from sipXpark relied on it, as media queues are usually full.

The queue can be limited or unlimited, the default is limited. Components
have the possibility to chose which kind of queue to use. As for this
moment the only component that is using a limited queue is sipXpark.
The other components are using an unlimited queue.

pprer